Use next character from before completion is invoked to avoid race
Fixes #672. This was a race where if the user typed more characters
after we sent a request for completions, then the range we used was bad
(Code compensates for this during insertion but we did not). To fix
this, just take the character before we wait for completions and use
that.