Replace all reference to .fsPath to a custom function
This uppercases the drive letters (Uri.fsPath lowercases). This seems to be what most things (including the analysis server) expect, so fixed a bunch of issues.

Could be very risky though if anything is relying on lowercase (for ex. if the analysis server is actually getting lowercase paths for some Windows users - hard to know without knowing where this comes from).