Add support for uncomment files, better timeout handling
createNonExistentMethod caused errors when working on Dart Code, so the source is now commented out and gets uncommented by the test (this is more convenient than putting all the Dart into the test, as when we're editing it we can uncomment temporarily to get syntax checking etc).

Also tweaked some timeouts and made a waitForNextAnalysis() that wraps calling an action in waiting for an analysis to complete.