Improve reuse of nodes across runs
This is more complicated than expected! Rather than blindly reusing a node based on its ID, it will require the ID to match *for the current run* otherwise it will do a name search (and include line number, to reduce false positives). If it finds a node, it will reuse it and update it to match the current run.

To complicate things, old nodes may still try and look up their parents by the old ID, so SuiteData's maps now contain dupes based on the old and new keys.

TODO: Refactor all of this crazy logic out of test_view into its own class, so that consuming the test results is much simpler and disconnected from the view!

Fixes #1108.