Pre-fill name upon creating a new project
Now asks for the project folder first, so it can determine a unique project name for creation should the pre-filled name be accepted within the corresponding dialog.

Project name determination only occurs for the initial, pre-filled name suggestion and will not apply if a different name is entered entirely.

Fixes #2989.